+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 # Changelog
 
+## [0.1.2](https://github.com/fpsvogel/pipeful/releases/tag/0.1.2)
+
+- Added an example reading log, `read.csv`.
+- Cleaned up `config.rb` and `load_items_test.rb`.
+
 ## [0.1.1](https://github.com/fpsvogel/pipeful/releases/tag/0.1.1)
 
 - Wrote tests for bad input.
-- Added the "help" command.
+- Added the `help` command.
 
 ## [0.1.0](https://github.com/fpsvogel/pipeful/releases/tag/0.1.0)